Add shodan search count before actually querying with net: filter
Official documentation says that credits are deducted on search query
filter usage and when additional pages with results are viewed. When enumerating
lots of AS networks it is usual to see shodan responses with 0 hosts in it, so
credits just getting wasted.

Although I didn't notice credits deduction, I decided to prepare drone for this.
Request to /shodan/host/search firstly is available to avoid that cases, but it
leads us to make x2 requests for every subnet. To address this slow
down 10 goroutines will be started for nets to ips processing.

Main function was a little reorganized to separate actually hosts quering from
getting initial list of hosts which is now in gatherIPsToSearch function. It
could be expanded in future to handle not only IPs and subnets but also domain names.